+++ src/transaction.cpp
@@ -10,7 +10,7 @@ int TraxManager::clear_trax() {
     return 0;
 }
 
-int TraxManager::rollback_trax(int tstamp) {
+int TraxManager::rollback_trax(unsigned long tstamp) {
     rnum.roll_back(tstamp);
     record.roll_back(tstamp);
     pending.roll_back(tstamp);
